+++ src/main/java/dev/redstudio/alfheim/asm/AlfheimPlugin.java
@@ -2,10 +2,14 @@ package dev.redstudio.alfheim.asm;
 
 import com.google.common.collect.ImmutableList;
 import net.minecraftforge.fml.relauncher.IFMLLoadingPlugin;
+import zone.rong.mixinbooter.Context;
 import zone.rong.mixinbooter.IEarlyMixinLoader;
+import zone.rong.mixinbooter.IMixinConfigHijacker;
 
+import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
 import static dev.redstudio.alfheim.ProjectConstants.ID;
 import static dev.redstudio.alfheim.ProjectConstants.LOGGER;
@@ -17,7 +21,7 @@ import static dev.redstudio.alfheim.ProjectConstants.NAME;
  */
 @IFMLLoadingPlugin.MCVersion("1.12.2")
 @IFMLLoadingPlugin.TransformerExclusions("dev.redstudio." + ID + ".asm")
-public final class AlfheimPlugin implements IFMLLoadingPlugin, IEarlyMixinLoader {
+public final class AlfheimPlugin implements IFMLLoadingPlugin, IEarlyMixinLoader, IMixinConfigHijacker {
 
     @Override
     public String[] getASMTransformerClass() {
@@ -45,13 +49,13 @@ public final class AlfheimPlugin implements IFMLLoadingPlugin, IEarlyMixinLoader
 
     @Override
     public List<String> getMixinConfigs() {
-        return ImmutableList.of("mixins." + ID +".json");
+        return ImmutableList.of("mixins." + ID + ".json");
     }
 
     @Override
     public boolean shouldMixinConfigQueue(final String mixinConfig) {
         switch (mixinConfig) {
-            case "mixins." + ID +".json":
+            case "mixins." + ID + ".json":
                 return !isCubicChunksInstalled();
             default:
                 return true;
@@ -72,5 +76,19 @@ public final class AlfheimPlugin implements IFMLLoadingPlugin, IEarlyMixinLoader
 
         LOGGER.warn("Cubic Chunks was detected, it uses it's own lighting engine, {} will not load", NAME);
         return true;
+    }
+
+    @Override
+    public Set<String> getHijackedMixinConfigs() {
+        return Collections.singleton("mixins.phosphor.json");
+    }
+
+    @Override
+    public Set<String> getHijackedMixinConfigs(final Context context) {
+        if (!context.isModPresent("phosphor-lighting"))
+            return Collections.emptySet();
+
+        LOGGER.warn("Phosphor/Hesperus was detected, {} replaces them entirely, will now attempt hijacking, you should remove Phosphor/Hesperus", NAME);
+        return getHijackedMixinConfigs();
     }
 }